Replaced ServerConfig / Builder with concrete implementations.
