+++ src/test/java/com/appsflyer/donkey/TestUtil.java
@@ -1,24 +1,25 @@
 /*
  * Copyright 2020 AppsFlyer
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
 package com.appsflyer.donkey;
 
 import clojure.lang.IPersistentMap;
 import com.appsflyer.donkey.server.ServerConfig;
-import com.appsflyer.donkey.server.ServerConfigBuilder;
+import com.appsflyer.donkey.server.ServerConfig.ServerConfigBuilder;
 import com.appsflyer.donkey.server.route.AbstractRouteCreator;
 import com.appsflyer.donkey.server.route.RouteCreator;
 import com.appsflyer.donkey.server.route.RouteDefinition;
