+++ src/test/java/com/appsflyer/donkey/server/handler/ContentTypeHandlerTest.java
@@ -1,17 +1,18 @@
 /*
  * Copyright 2020 AppsFlyer
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
 package com.appsflyer.donkey.server.handler;
@@ -19,7 +20,7 @@ package com.appsflyer.donkey.server.handler;
 import com.appsflyer.donkey.TestUtil;
 import com.appsflyer.donkey.server.Server;
 import com.appsflyer.donkey.server.ServerConfig;
-import com.appsflyer.donkey.server.ServerConfigBuilder;
+import com.appsflyer.donkey.server.ServerConfig.ServerConfigBuilder;
 import com.appsflyer.donkey.server.exception.ServerInitializationException;
 import com.appsflyer.donkey.server.exception.ServerShutdownException;
 import com.appsflyer.donkey.server.route.RouteDefinition;
